+++ receiver/prometheusreceiver/internal/otlp_metricsbuilder.go
@@ -59,25 +59,28 @@ func getBoundary(metricType pmetric.MetricDataType, labels labels.Labels) (float
 	return strconv.ParseFloat(v, 64)
 }
 
-func convToMetricType(metricType textparse.MetricType) pmetric.MetricDataType {
+// convToPdataMetricType returns the data type and if it is monotonic
+func convToMetricType(metricType textparse.MetricType) (pmetric.MetricDataType, bool) {
 	switch metricType {
 	case textparse.MetricTypeCounter:
 		// always use float64, as it's the internal data type used in prometheus
-		return pmetric.MetricDataTypeSum
+		return pmetric.MetricDataTypeSum, true
 	// textparse.MetricTypeUnknown is converted to gauge by default to prevent Prometheus untyped metrics from being dropped
 	case textparse.MetricTypeGauge, textparse.MetricTypeUnknown:
-		return pmetric.MetricDataTypeGauge
+		return pmetric.MetricDataTypeGauge, false
 	case textparse.MetricTypeHistogram:
-		return pmetric.MetricDataTypeHistogram
+		return pmetric.MetricDataTypeHistogram, true
 	// dropping support for gaugehistogram for now until we have an official spec of its implementation
 	// a draft can be found in: https://docs.google.com/document/d/1KwV0mAXwwbvvifBvDKH_LU1YjyXE_wxCkHNoCGq1GX0/edit#heading=h.1cvzqd4ksd23
 	// case textparse.MetricTypeGaugeHistogram:
 	//	return <pdata gauge histogram type>
 	case textparse.MetricTypeSummary:
-		return pmetric.MetricDataTypeSummary
+		return pmetric.MetricDataTypeSummary, true
+	case textparse.MetricTypeInfo, textparse.MetricTypeStateset:
+		return pmetric.MetricDataTypeSum, false
 	default:
-		// including: textparse.MetricTypeGaugeHistogram, textparse.MetricTypeInfo, textparse.MetricTypeStateset
-		return pmetric.MetricDataTypeNone
+		// including: textparse.MetricTypeGaugeHistogram
+		return pmetric.MetricDataTypeNone, false
 	}
 }